Partial Build - Copy/Paste
Added the ability to copy selections (Ctrl +C, or edit menu) or a color
combination (Ctrl + Alt + C, or menu) to the clipboard.
Color combinations can be pasted (Ctrl + V) onto all selected rooms.
Copied selections can not yet be pasted.

Because this uses clipboard it can easily be pasted across different
maps.

This function has not been thoroughly tested yet and there is likely to
be a higher presence of bugs.