Copy and Paste fully functional
This build can now copy and paste rooms, connections and color schemes.

Copying fully docked lines without at least one of the rooms it is
connected to can be troublesome as there's no point of reference copied.
Either it won't paste or it will paste as a dot in the centre of the
map.

Copying lines without any rooms will cause the lines to paste in the
centre of the map.