Added UNDO handling to automap
When an UNDO is performed, the resulting room name is printed, followed
by "[Previous turn undone.]" I added simple code to ignore this line
when processing paragraphs, which makes this room look like one with a
non-verbose description. It's not perfect (for that, you'd have to store
a state graph), but it should handle most cases.