Disable call to user32.dll:SendMessage().
This call causes disruptive exceptions when double clicking on a room.
It's not immediately obvious to me what this does.  It appears to be
related to what I would call tooltips, which are not currently being
displayed in Linux when hovering over a room.

Perhaps in the future this should be replaced with something else.
For now my goal is to suppress annoying behavior wher Trizbort prompts
the user to save the file everytime the room dialog is opened.
Presumably because an exception occurred?