Remove the NUM and CAPS widgets from the scrollbar.
These widgets crash the Linux Mono version.  I thought that there
might be a more portable way to handle at least querying the status of
the CapsLock, NumLock, and ScrollLock Keys, but it turns out that the
API just isn't implemented in Mono yet.  See:

https://github.com/mono/mono/blob/main/mcs/class/System.Windows.Forms/System.Windows.Forms/Control.cs#L3487

Fortunately the Trizbort feature isn't terribly important, and the
goal for users on other platforms would be to avoid crashes by
accidentally clicking on the widgets, which is what this commit
accomplishes by removing the widgets from the status bar entirely.