Possible fix for the Door icons crash.
I'm not 100% confident that I really understand the problem here, but
I tracked it back to this call to DrawImage, and the exception seems
to indicate that there was a "busy" problem presumably related to Mono
using System.Drawing to render the Windows forms.

On a whim, I wondered if an XImage would behave differently, and it
did.  No crash, and the icons seem to be drawn correctly.  Perhaps
there's a better fix than this for the issue, but this gets it working
at least.  If someone else can find a better fix or further explain
this one, that feedback would be welcome.