Object positioning check boxes experiment #2.
This commit also displays _very_ poorly in Mono on Linux, but I've
included it to ensure that the problem in the previous commit is not
related to the encoding of the source code file.  For me, both commits
produce essentially identical results (correct on console, nothing on
the check boxes inside the application).