Object positioning check boxes experiment #3.
This commit is an improvement on Linux if you don't have the Wingdings
font (which is provavly all Linux users by default).  Without the
Wingdings font, you get random characters that don't communicate the
purpose of the checkbox widgets at all.  With this commit, the north,
south, east and west arrows render reasonably and the other five
characters render as "missing character squares" which in my opinion
makes it a little more obvious what was intended.

Unfortunately, despite being only a few characters away from each
other in unicode, on my system the cardinal directions are "normal
width" unicode characters rendered at with a lighter weight, while the
five missing characters are extra wide, and substantially heavier in
their rendering.  So they're definitely not "harmonious".

If it works (and looks good) in Windows, I think this solution should
be considered temporarily since it's an improvement on Linux, and it
might get fixed before the commits #1 or #2, which is what should
really be done long term. If this doesn't look good on Windows, we
should revert to the original source code, and advise non-Windows
users to somehow obtain and install the Wingdings font.