This catch was hiding real failures in Mono.
This catch clause was why text was mysteriously missing without there
being any errors.  Unfortunately, with it disabled, simply creating a
room causes Trizbort to crash.  Perhaps there's a good reason for this
catch on Windows, but at least for testing Mono on Linux, it should
probably be disabled.  It might also be a good idea to report that
something went wrong in the future, but for now I'd like actual
exceptions if possible.