Get back to a place where you can add a room in Mono.
So this commit shows where the problems revealed by the previous
commit really are.  It's apparently all about graphics.MeasureString
and graphics.DrawString.  There are some other places in the code that
use these same methods, and they probably crash too, but I haven't
attempted to disable them because they don't interfere with adding
rooms or common functionality.

I was surprised to learn that the problem can be triggered by
converting the Font to a XFont.  Perhaps the MeasureString and
DrawString calls would actually work if there was a valid XFont?

This commit makes it clearer to me why PDFSharp is fundamentally part
of the problem, and why a correct Mono port might need to drop support
for PDFSharp entirely.  The traceback generated by the initialXFont
line is mentionded first in 2017, and appears to part of a long
running thread about PDFSharp:

https://github.com/empira/PDFsharp/issues/9 in 2017
https://github.com/empira/PDFsharp/issues/66 in 2018
https://github.com/empira/PDFsharp/issues/136